인터페이스 분리 1차작업 및 불필요한 임포트 및 경고 코드 정리
