+++ auth-service/src/test/java/com/playdata/authservice/AuthServiceImplApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class AuthServiceApplicationTests {
+class AuthServiceImplApplicationTests {
 
     @Test
     void contextLoads() {
